Add a factory method
