+++ jrpg/models/demons/book.py
@@ -1,8 +1,3 @@
-from models.demons.chapters.katakana import Demon_chapter_katakana
-from models.demons.chapters.hiragana import Demon_chapter_hiragana
-from models.demons.chapters.kanaword import Demon_chapter_kanaword
-from models.demons.chapters.trad import Demon_chapter_trad
-from models.demons.chapters.kanji import Demon_chapter_kanji
 
 from random import choice
 
@@ -24,37 +19,13 @@ class Book_of_demons:
         self.demons[chapter.get_title_of_the_chapter()] = []
         self.demons[chapter.get_title_of_the_chapter()] = chapter.get_the_list_of_demon()
 
-    # Pseudo factory method
-    # TODO create a specific class ?
-    # there is probably a python "way" to do it
-    def get_chapter_for_book_of_demon(self, type, filename):
-        if type == 'katakana':
-            return Demon_chapter_katakana(filename)
-        elif type == 'hiragana':
-            return Demon_chapter_hiragana(filename)
-        elif type == 'kanaword':
-            return Demon_chapter_kanaword(filename)
-        elif type == 'traduction':
-            return Demon_chapter_trad(filename)
-        elif type == 'kanji':
-            return Demon_chapter_kanji(filename)
 
-    def __init__(self, config={}):
+    def __init__(self, chapter_factory,  config={}):
         self.demons = {}
-        # TODO give a dictionnary {'katakana': 'data/demonkatakana.txt' ..}.
-        # and do a foreach in this dictionnary
-        self.add_chapter_to_the_book(
-                self.get_chapter_for_book_of_demon(
-                    'katakana',
-                    'data/demons-katakana.txt'))
-        self.add_chapter_to_the_book(self.get_chapter_for_book_of_demon('hiragana',
-            'data/demons-hiragana.txt'))
-        self.add_chapter_to_the_book(self.get_chapter_for_book_of_demon('kanaword',
-            'data/demons-kanawords.txt'))
-        self.add_chapter_to_the_book(self.get_chapter_for_book_of_demon('traduction',
-            'data/demons-kanawords.txt'))
-        self.add_chapter_to_the_book(self.get_chapter_for_book_of_demon('kanji',
-            'data/demons-kanji.txt'))
+        for key, filename in config.items():
+            self.add_chapter_to_the_book(
+                    chapter_factory.get_chapter_for_book_of_demon(key, filename))
+
 
         #Debug:
         #for chapter_id in self.demons:
